refactor(k8s-client): preserve AutoDetect in-cluster semantics; extract FakeCrdModel fixture
- AutoDetect probes in-cluster availability without the caller namespace, matching the original
  AutoDetectClientFacadeFactory (the in-cluster branch also requires the namespace file).
- Move FakeCrdModel to its own PSR-4 file so KubernetesApiClientFacadeFactoryTest no longer depends
  on a fixture defined inside another test file (fixed an isolation-order error).
